requirements.txt
aiohttp==3.7.4
async-timeout==4.0.2
charset-normalizer==2.0.4
multidict==5.1.0
pandas==1.3.3
python-socketio[client]==5.5.1
